Give the formation geometry its own home
Both halves of the doctrine ask the same questions about a formation:
where its centre is and how wide it may be. That arithmetic sat in
MutualSupportDeployment, so the movement half called into a class named
for deployment. The dependency ran the wrong way, and neither class was
named for what it held.

Moves formationRadius, formationRadiusFor, weightedCentroid, centroid and
MINIMUM_SPACING_HEXES into FormationGeometry. Deployment keeps the
ordering of candidate hexes; the ranker keeps judging paths; both now
depend on the shared shape rather than on each other.

No behaviour change. The methods moved unaltered and every caller was
repointed.

  MutualSupportDeployment  -> FormationGeometry
  MutualSupportPathRanker  -> FormationGeometry, FormationSide, SupportEnvelope
  FormationGeometry        -> SupportEnvelope
  FormationSide            -> nothing
  SupportEnvelope          -> nothing

MutualSupportDeployment drops from 326 lines to 213.
